+++ src/test/java/org/folio/qm/converter/MarcQmToMarcRecordConverterTest.java
@@ -45,7 +45,7 @@ class MarcQmToMarcRecordConverterTest {
     var source = new BaseMarcRecord().leader("test");
     var tag = "245";
 
-    when(fieldsConverter.convertQmFields(any(), any())).thenReturn(List.of(new DataFieldImpl(tag, ' ', ' ')));
+    when(fieldsConverter.convertQmFields(any(), any())).thenReturn(List.of(new DataFieldImpl(tag, ' ', '\\')));
 
     var actual = converter.convert(source);
 
@@ -54,5 +54,21 @@ class MarcQmToMarcRecordConverterTest {
     assertThat(actual.dataFields().get(0).indicators()).isNotNull();
     assertThat(actual.dataFields().get(0).indicators()).hasSize(2);
     actual.dataFields().get(0).indicators().forEach(ind -> assertEquals('#', ind.value()));
+  }
+
+  @Test
+  void convert_invalidFieldIndicatorContent() {
+    var source = new BaseMarcRecord().leader("test");
+    var tag = "245";
+
+    when(fieldsConverter.convertQmFields(any(), any())).thenReturn(List.of(new DataFieldImpl(tag, '#', '#')));
+
+    var actual = converter.convert(source);
+
+    assertThat(actual).isNotNull();
+    assertThat(actual.dataFields()).hasSize(1);
+    assertThat(actual.dataFields().get(0).indicators()).isNotNull();
+    assertThat(actual.dataFields().get(0).indicators()).hasSize(2);
+    actual.dataFields().get(0).indicators().forEach(ind -> assertEquals('\\', ind.value()));
   }
 }